move menu kind to menu module
