+++ src/packages/core/menu/section-sidebar-menu-with-entity-actions/index.ts
