+++ src/packages/core/menu/section-sidebar-menu/section-sidebar-menu.element.ts
@@ -1,4 +1,4 @@
-import type { ManifestSectionSidebarAppBaseMenu, ManifestSectionSidebarAppMenuKind } from '../types.js';
+import type { ManifestSectionSidebarAppBaseMenu, ManifestSectionSidebarAppMenuKind } from './types.js';
 import { UmbTextStyles } from '@umbraco-cms/backoffice/style';
 import { css, html, customElement, property } from '@umbraco-cms/backoffice/external/lit';
 import type { ManifestMenu, UmbExtensionManifestKind } from '@umbraco-cms/backoffice/extension-registry';
