+++ src/packages/core/menu/section-sidebar-menu/types.ts
@@ -1,0 +1,26 @@
+import type { ManifestSectionSidebarApp } from '@umbraco-cms/backoffice/section';
+
+export interface MetaSectionSidebarAppMenuKind {
+	label: string;
+	menu: string;
+}
+
+export interface ManifestSectionSidebarAppBaseMenu extends ManifestSectionSidebarApp {
+	type: 'sectionSidebarApp';
+	meta: MetaSectionSidebarAppMenuKind;
+}
+
+export interface ManifestSectionSidebarAppMenuKind extends ManifestSectionSidebarAppBaseMenu {
+	type: 'sectionSidebarApp';
+	kind: 'menu';
+}
+
+export interface ManifestSectionSidebarAppMenuWithEntityActionsKind extends ManifestSectionSidebarAppBaseMenu {
+	type: 'sectionSidebarApp';
+	kind: 'menuWithEntityActions';
+	meta: MetaSectionSidebarAppMenuWithEntityActionsKind;
+}
+
+export interface MetaSectionSidebarAppMenuWithEntityActionsKind extends MetaSectionSidebarAppMenuKind {
+	entityType: string;
+}
