+++ src/packages/core/section/extensions/index.ts
@@ -1,5 +1,5 @@
 import type { ManifestSectionRoute } from './section-route.extension.js';
-import type { ManifestSectionSidebarApp, ManifestSectionSidebarAppMenuKind } from './section-sidebar-app.extension.js';
+import type { ManifestSectionSidebarApp } from './section-sidebar-app.extension.js';
 import type { ManifestSectionView } from './section-view.extension.js';
 import type { ManifestSection } from './section.extension.js';
 
@@ -10,12 +10,7 @@ export type * from './section.extension.js';
 
 export type * from './types.js';
 
-type UmbSectionExtensions =
-	| ManifestSection
-	| ManifestSectionRoute
-	| ManifestSectionSidebarApp
-	| ManifestSectionSidebarAppMenuKind
-	| ManifestSectionView;
+type UmbSectionExtensions = ManifestSection | ManifestSectionRoute | ManifestSectionSidebarApp | ManifestSectionView;
 
 declare global {
 	interface UmbExtensionManifestMap {
