+++ src/packages/core/section/extensions/section-sidebar-app.extension.ts
@@ -6,28 +6,3 @@ export interface ManifestSectionSidebarApp
 		ManifestWithDynamicConditions<UmbExtensionCondition> {
 	type: 'sectionSidebarApp';
 }
-
-export interface ManifestSectionSidebarAppBaseMenu extends ManifestSectionSidebarApp {
-	type: 'sectionSidebarApp';
-	meta: MetaSectionSidebarAppMenuKind;
-}
-
-export interface ManifestSectionSidebarAppMenuKind extends ManifestSectionSidebarAppBaseMenu {
-	type: 'sectionSidebarApp';
-	kind: 'menu';
-}
-
-export interface MetaSectionSidebarAppMenuKind {
-	label: string;
-	menu: string;
-}
-
-export interface ManifestSectionSidebarAppMenuWithEntityActionsKind extends ManifestSectionSidebarAppBaseMenu {
-	type: 'sectionSidebarApp';
-	kind: 'menuWithEntityActions';
-	meta: MetaSectionSidebarAppMenuWithEntityActionsKind;
-}
-
-export interface MetaSectionSidebarAppMenuWithEntityActionsKind extends MetaSectionSidebarAppMenuKind {
-	entityType: string;
-}
