+++ src/packages/core/section/index.ts
@@ -4,8 +4,8 @@ export * from './section-default.element.js';
 export * from './section-main/index.js';
 export * from './section-picker-modal/section-picker-modal.token.js';
 export * from './section-sidebar-context-menu/index.js';
-export * from './section-sidebar-menu-with-entity-actions/index.js';
-export * from './section-sidebar-menu/index.js';
+export * from '../menu/section-sidebar-menu-with-entity-actions/index.js';
+export * from '../menu/section-sidebar-menu/index.js';
 export * from './section-sidebar/index.js';
 export * from './section.context.js';
 export type * from './types.js';
